server/pyproject.toml
[project]
name = "communityenergylabsserver"
version = "0.1.0"
description = "Take home assessment for CommunityEnergyLabs"
authors = [
    {name = "Kenny LaMarca",email = "kennywu1005@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=8.3.4,<9.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "flask-cors (>=5.0.0,<6.0.0)",
    "marshmallow (>=3.26.0,<4.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "flask-caching (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
